+++ packages/server/src/tests/banners/propensityModelTest/propensityModelTestData.ts
@@ -1,33 +1,23 @@
-import { BannerContent, SecondaryCtaType } from '@sdc/shared/types';
+import { logInfo } from '../../../utils/logging';
+import { streamS3DataByLine } from '../../../utils/S3';
+import { isProd } from '../../../lib/env';
 
-export const USDigisubContent: BannerContent = {
-    heading: 'Power open, independent journalism',
-    paragraphs: [
-        'Millions turn to the Guardian every day for fiercely independent journalism that’s open and free for all. We have no shareholders, no billionaire owner and no commercial or political bosses. Just the passion and determination to bring readers quality, truth-seeking reporting on the world, its people and power. This makes us different. Show your support today by becoming a digital subscriber, from just $2.50 a week. Doing so helps to protect our vital independence, it keeps us free of a paywall, and makes a real difference for our future.',
-    ],
-    highlightedText:
-        'And to say thank you, we’ll give you ad-free reading, and exclusive access to premium features on our award-winning apps.',
-    cta: {
-        baseUrl:
-            'https://support.theguardian.com/subscribe/digital/checkout?promoCode=DK0NT24WG&period=Monthly',
-        text: 'Subscribe',
-    },
-    secondaryCta: {
-        type: SecondaryCtaType.Custom,
-        cta: {
-            baseUrl: 'https://support.theguardian.com/subscribe/digital',
-            text: 'Find out more',
+const guardianWeeklyHighPropensityIds: Set<string> = new Set<string>();
+const fetchHighPropensityIds = (): void => {
+    logInfo('Loading guardianWeeklyHighPropensityIds...');
+    streamS3DataByLine(
+        'support-admin-console',
+        `${isProd ? 'PROD' : 'CODE'}/guardian-weekly-propensity-test/ids.txt`,
+        line => guardianWeeklyHighPropensityIds.add(line),
+        () => {
+            logInfo(
+                `Loaded ${guardianWeeklyHighPropensityIds.size} guardianWeeklyHighPropensityIds`,
+            );
         },
-    },
+    );
 };
 
-// TODO
-export const GWContent: BannerContent = {
-    heading: '',
-    paragraphs: [],
-    highlightedText: '',
-    cta: {
-        baseUrl: '',
-        text: '',
-    },
-};
+fetchHighPropensityIds();
+
+export const isInPropensityTest = (browserId: string): boolean =>
+    guardianWeeklyHighPropensityIds.has(browserId);
